+++ sql-plugin/src/main/330db/scala/com/nvidia/spark/rapids/shims/SparkShims.scala
@@ -24,7 +24,7 @@ import org.apache.spark.sql.catalyst.trees.TreePattern._
 import org.apache.spark.sql.execution.datasources._
 import org.apache.spark.sql.execution.datasources.parquet.ParquetFilters
 
-object SparkShimImpl extends Spark321PlusDBShims {
+object SparkShimImpl extends Spark321PlusDBShims with RoundingShims {
   // AnsiCast is removed from Spark3.4.0
   override def ansiCastRule: ExprRule[_ <: Expression] = null
 
@@ -43,6 +43,9 @@ object SparkShimImpl extends Spark321PlusDBShims {
     new ParquetFilters(schema, pushDownDate, pushDownTimestamp, pushDownDecimal, pushDownStartWith,
       pushDownInFilterThreshold, caseSensitive, datetimeRebaseMode)
   }
+
+    override def getExprs: Map[Class[_ <: Expression], ExprRule[_ <: Expression]] =
+      super.getExprs ++ roundingExprs
 }
 
 trait ShimExtractValue extends ExtractValue {
